refactor: 컴포넌트 분리 및 isSearchView 상태값 recoil로 변경
